OJORISE-45 modify the file structure
