+++ src/app/login/page.tsx
@@ -1,4 +1,4 @@
-import LoginPage from "@/components/page-component/LoginPage";
+import LoginPage from "@/components/page-component/login/LoginPage";
 
 export default function LoginPageRoute() {
   return <LoginPage />;
