+++ src/app/login/success/page.tsx
@@ -1,4 +1,4 @@
-import SuccessPage from "@/components/page-component/SuccessPage";
+import SuccessPage from "@/components/page-component/success/SuccessPage";
 
 export default function SuccessPageRoute() {
   return <SuccessPage/>;
