+++ src/app/main-page/page.tsx
@@ -1,4 +1,4 @@
-import MainPage from "@/components/page-component/MainPage";
+import MainPage from "@/components/page-component/main-page/MainPage";
 
 export default function Home() {
   return <MainPage />;
