+++ src/app/signup/page.tsx
@@ -1,4 +1,4 @@
-import SignupPage from "@/components/page-component/SignupPage";
+import SignupPage from "@/components/page-component/signup/SignupPage";
 
 export default function SignupPageRoute() {
   return <SignupPage />;
