+++ src/components/page-component/main-page/recommanded-list/CompareModeToggle.tsx
