+++ src/components/page-component/main-page/recommanded-list/PlanCard.tsx
