+++ src/components/page-component/main-page/recommanded-list/SelectedPlanViewer.tsx
