+++ src/components/page-component/signup/SignupPage.tsx
@@ -1,4 +1,4 @@
-import VerticalLinearStepper from "../common/VerticalLinearStepper";
+import VerticalLinearStepper from "./VerticalLinearStepper";
 export default function SignupPage() {
   return (
     <div className="flex min-h-svh flex-col items-center justify-center p-6 md:p-10">
